Substitui SwipeableDrawer por TemporaryDrawer
